# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "HS-SPORT API",
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/player/player-standings": {
            "post": {
                "tags": [
                    "Player"
                ],
                "summary": "Player Standings",
                "description": "Player Standings",
                "operationId": "api.player.player.standings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "playerId",
                        "in": "query",
                        "description": "player ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "viewType",
                        "in": "query",
                        "description": "state || school || team",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/statistics/leaderboard": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Leaderboard",
                "description": "Stats - Leaderboard",
                "operationId": "api.statistics.leaderboard",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "1 or 2",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "1 or 2 or null",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "varsityId",
                        "in": "query",
                        "description": "varsity ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "team ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "stateId",
                        "in": "query",
                        "description": "state ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "city ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/statistics/scoreboard": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Scoreboard",
                "description": "Stats - Scoreboard",
                "operationId": "api.statistics.scoreboard",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "team ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "stateId",
                        "in": "query",
                        "description": "state ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "cityId",
                        "in": "query",
                        "description": "city ID",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/statistics/sport-scoreboard": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Sport Scoreboard",
                "description": "Stats - Sport Scoreboard",
                "operationId": "api.statistics.sport-scoreboard",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "gender ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "teamType",
                        "in": "query",
                        "description": "team TYPE",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/statistics/sport-leaderboard": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Sport Leaderboard",
                "description": "Stats - Sport Leaderboard",
                "operationId": "api.statistics.sport-leaderboard",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "gender ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "teamType",
                        "in": "query",
                        "description": "team TYPE",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "withoutTypes",
                        "in": "query",
                        "description": "without TYPE",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/statistics/sport-team-standings": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Sport Team Standings",
                "description": "Stats - Sport Team Standings",
                "operationId": "api.statistics.sport-team-standings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sportId",
                        "in": "query",
                        "description": "sport ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "seasonId",
                        "in": "query",
                        "description": "season ID",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "genderId",
                        "in": "query",
                        "description": "gender ID",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "teamType",
                        "in": "query",
                        "description": "team TYPE",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "withoutTypes",
                        "in": "query",
                        "description": "without TYPE",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/statistics/scoreboard-game-detail": {
            "post": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Stats - Scoreboard detail",
                "description": "Stats - Scoreboard detail",
                "operationId": "api.statistics.scoreboard-game-detail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "query",
                        "description": "game ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Show error message"
                    },
                    "401": {
                        "description": "Redirect to login page"
                    },
                    "409": {
                        "description": "Redirect to page 404"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}